requirements.txt
aniso8601==10.0.0
attrs==25.1.0
blinker==1.9.0
CacheControl==0.14.2
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
firebase-admin==6.6.0
Flask==3.1.0
Flask-Cors==5.0.0
flask-restx==1.3.0
Flask-Testing==0.8.1
gcloud==0.18.3
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-secret-manager==2.23.0
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.67.0
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
importlib_resources==6.5.2
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwcrypto==1.5.6
MarkupSafe==3.0.2
msgpack==1.1.0
oauth2client==4.1.3
packaging==24.2
pluggy==1.5.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.10.1
pyparsing==3.2.1
Pyrebase4==4.8.0
pytest==8.3.4
python-dotenv==1.0.1
python-jwt==4.1.0
pytz==2025.1
referencing==0.36.2
requests==2.32.3
requests-toolbelt==0.10.1
rpds-py==0.22.3
rsa==4.9
setuptools==75.8.0
six==1.17.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.20
Werkzeug==3.1.3
